+++ example/android/app/src/main/kotlin/com/wid_mage/example/example/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.wid_mage.example.example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}