+++ lib/src/util/wid_mage.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:wid_mage/src/util/wid_mage_controller.dart';
 
 class WidMage extends StatelessWidget {